cleanups to scripts and READMEs
