+++ rust-runtime/aws-smithy-client/README.md
@@ -1,6 +1,8 @@
 # aws-smithy-client
 
-`aws-smithy-client` defines a Tower-based client that implements functionality that exists across all smithy-based service clients including:
+`aws-smithy-client` defines a Tower-based client that implements functionality that exists across all service clients
+generated by [smithy-rs](https://github.com/awslabs/smithy-rs) including:
+
 - Retries
 - Connector, request attempt, and multi-request timeouts
 - Configurable middleware
