+++ rust-runtime/aws-smithy-http/README.md
@@ -1,6 +1,6 @@
 # aws-smithy-http
 
-Core HTTP primitives for Smithy-based service clients including:
+Core HTTP primitives for service clients generated by [smithy-rs](https://github.com/awslabs/smithy-rs) including:
 - HTTP Body implementation
 - Endpoint support
 - HTTP header deserialization
