+++ rust-runtime/aws-smithy-json/README.md
@@ -1,6 +1,6 @@
 # aws-smithy-json
 
-JSON serialization and deserialization primitives for smithy-rs generated services.
+JSON serialization and deserialization primitives for clients and servers generated by [smithy-rs](https://github.com/awslabs/smithy-rs).
 
 <!-- anchor_start:footer -->
 This crate is part of the [AWS SDK for Rust](https://awslabs.github.io/aws-sdk-rust/) and the [smithy-rs](https://github.com/awslabs/smithy-rs) code generator. In most cases, it should not be used directly.
