+++ rust-runtime/aws-smithy-xml/README.md
@@ -1,6 +1,6 @@
 # aws-smithy-xml
 
-XML serialization and deserialization primitives for Smithy-based services.
+XML serialization and deserialization primitives for services generated by [smithy-rs](https://github.com/awslabs/smithy-rs).
 
 <!-- anchor_start:footer -->
 This crate is part of the [AWS SDK for Rust](https://awslabs.github.io/aws-sdk-rust/) and the [smithy-rs](https://github.com/awslabs/smithy-rs) code generator. In most cases, it should not be used directly.
